# docs/swagger.yaml
basePath: /api
definitions:
  msg.Message:
    properties:
      code:
        $ref: '#/definitions/msg.RespCode'
        type: object
      data:
        type: object
      error:
        type: string
    type: object
  msg.RespCode:
    type: object
  user.User:
    properties:
      createdAt:
        type: string
      createdIP:
        type: string
      id:
        type: integer
      password:
        type: string
      updatedAt:
        type: string
      userAgent:
        type: string
      username:
        type: string
    type: object
  user.loginReq:
    $ref: '#/definitions/user.registerReq'
  user.registerReq:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  description: This is a sample server celler server.
  license: {}
  title: APP Demo API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: login user
      parameters:
      - description: login info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/user.loginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.Message'
      summary: login user
      tags:
      - User
  /register:
    post:
      consumes:
      - application/json
      description: register user
      parameters:
      - description: register info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/user.registerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.Message'
      summary: register user
      tags:
      - User
  /userinfo:
    get:
      consumes:
      - application/json
      description: user info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
      summary: user info
      tags:
      - User
swagger: "2.0"